requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
streamlit
pydantic
